requirements.txt
aioredis==1.3.1
amqp==2.6.1
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.1.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
billiard==3.6.3.0
braintree==4.3.0
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
certifi==2020.6.20
cffi==1.14.2
channels==2.4.0
channels-redis==3.0.1
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==3.0
cssselect2==0.3.0
cycler==0.10.0
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1
django-allauth==0.42.0
django-autoslug==1.9.8
django-bootstrap3==14.1.0
django-braces==1.14.0
django-classy-tags==1.0.0
django-cms==3.7.4
django-countries==6.1.3
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-embed-video==1.3.3
django-extensions==3.0.5
django-filer==1.7.1
django-formtools==2.2
django-heroku==0.3.1
django-js-asset==1.2.2
django-localflavor==3.0.1
django-memcache-status==2.2
django-model-utils==4.0.0
django-mptt==0.11.0
django-parler==2.1
django-polymorphic==2.1.2
django-registration-redux==2.8
django-rosetta==0.9.4
django-sekizai==1.1.0
django-taggit==1.3.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangocms-admin-style==1.5.0
djangocms-attributes-field==1.2.0
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==1.5.0
djangocms-installer==1.2.3
djangocms-link==2.6.1
djangocms-picture==2.4.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
djangorestframework==3.11.1
easy-thumbnails==2.7
flower==0.9.5
gunicorn==20.0.4
hiredis==1.1.0
html5lib==1.1
humanize==2.6.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.7.0
incremental==17.5.0
isort==5.4.2
kiwisolver==1.2.0
kombu==4.6.11
lazy-object-proxy==1.4.0
Markdown==3.2.2
matplotlib==3.3.1
mccabe==0.6.1
msgpack==1.0.0
numpy==1.19.1
oauthlib==3.1.0
parse==1.16.0
pep8==1.7.1
Pillow==7.2.0
pip-autoremove==0.9.1
polib==1.1.0
prometheus-client==0.8.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pylint-django==2.3.0
pylint-plugin-utils==0.6
pyOpenSSL==19.1.0
pyparsing==2.4.7
Pyphen==0.9.5
python-dateutil==2.8.1
python-decouple==3.3
python-memcached==1.59
python-stdnum==1.14
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
stripe==2.51.0
tinycss2==1.0.2
toml==0.10.1
tornado==6.0.4
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
tzlocal==2.1
Unidecode==1.1.1
urllib3==1.25.10
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
